Optocoupler: Parse: Evaluate sign of deviation
With the recent hardware migration to my 3d printed case, the LED
is now driven by a resistor to prevent a short. In addition to that,
the PR is significantly further away from the LED than during the
breadboard setup. As a result, the PR reads significantly lower values
than before. A quick log has proven that the received signals result,
for the most part, in a lower value that than the brightness provided to
the LED. Interestingly, there's also a point where the LED input is lower to
the value received by the PR. As a result, we no longer only subtract the deviation,
but compare the LED brightness and PR input first, and then conclude whether to
subtract or add the deviation.

Signed-off-by: Patrick Pedersen <ctx.xda@gmail.com>